requirements.txt
asgiref==3.5.2
celery==5.2.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
geographiclib==1.52
geopy==2.2.0
factory-boy==3.2.1
Faker==15.1.0
kombu==5.2.4
pika==1.3.1
pika-stubs==0.1.3
psycopg2==2.9.4
idna==3.4
numpy==1.23.4
pandas==1.5.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.2
tzdata==2022.5
requests==2.28.1
six==1.16.0
urllib3==1.26.12
